fix: don't reset when changing blacklist items
ref #2318
